aobench/scenario-client/pyproject.toml
[project]
name = "scenario-client"
version = "0.1.0"
description = "scenario server client"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "load-dotenv>=0.1.0",
    "mlflow>=3.4.0",
]

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
